[WIP] UnityDoorstop 4 support
