+++ doorstop/doorstop_config_il2cpp.ini
@@ -1,18 +1,53 @@
-[UnityDoorstop]
-# Specifies whether assembly executing is enabled
-enabled=true
-# Specifies the path (absolute, or relative to the game's exe) to the DLL/EXE that should be executed by Doorstop
-targetAssembly=BepInEx\core\BepInEx.IL2CPP.dll
-# Specifies whether Unity's output log should be redirected to <current folder>\output_log.txt
-redirectOutputLog=false
+# General options for Unity Doorstop
+[General]
 
-[MonoBackend]
-runtimeLib=mono\MonoBleedingEdge\EmbedRuntime\mono-2.0-sgen.dll
-configDir=mono\MonoBleedingEdge\etc
-corlibDir=mono\Managed
-# Specifies whether the mono soft debugger is enabled
-debugEnabled=false
-# Specifies whether the mono soft debugger should suspend the process and wait for the remote debugger
-debugSuspend=false
-# Specifies the listening address the soft debugger
-debugAddress=127.0.0.1:10000
+# Enable Doorstop?
+enabled=true
+
+# Path to the assembly to load and execute
+# NOTE: The entrypoint must be of format `static void Doorstop.Entrypoint.Start()`
+target_assembly=BepInEx\core\BepInEx.IL2CPP.dll
+
+# If true, Unity's output log is redirected to <current folder>\output_log.txt
+redirect_output_log=false
+
+# If enabled, DOORSTOP_DISABLE env var value is ignored
+# USE THIS ONLY WHEN ASKED TO OR YOU KNOW WHAT THIS MEANS
+ignore_disable_switch=false
+
+
+# Options specific to running under Unity Mono runtime
+[UnityMono]
+
+# Overrides default Mono DLL search path
+# Sometimes it is needed to instruct Mono to seek its assemblies from a different path
+# (e.g. mscorlib is stripped in original game)
+# This option causes Mono to seek mscorlib and core libraries from a different folder before Managed
+# Original Managed folder is added as a secondary folder in the search path
+dll_search_path_override=
+
+# If true, Mono debugger server will be enabled
+debug_enabled=false
+
+# When debug_enabled is true, this option specifies whether Doorstop should initialize the debugger server
+# If you experience crashes when starting the debugger on debug UnityPlayer builds, try setting this to false
+debug_start_server=true
+
+# When debug_enabled is true, specifies the address to use for the debugger server
+debug_address=127.0.0.1:10000
+
+# If true and debug_enabled is true, Mono debugger server will suspend the game execution until a debugger is attached
+debug_suspend=false
+
+# Options sepcific to running under Il2Cpp runtime
+[Il2Cpp]
+
+# Path to coreclr.dll that contains the CoreCLR runtime
+# [MonoBackend]
+# runtimeLib=mono\MonoBleedingEdge\EmbedRuntime\mono-2.0-sgen.dll
+coreclr_path=
+
+# Path to the directory containing the managed core libraries for CoreCLR (mscorlib, System, etc.)
+# [MonoBackend]
+# corlibDir=mono\Managed
+corlib_dir=
